% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/influxdb.R
\name{influxdb_json2dataframe}
\alias{influxdb_json2dataframe}
\title{Transform influxdb json result into a data.frame}
\usage{
influxdb_json2dataframe(json)
}
\arguments{
\item{json}{raw influxdb json result}
}
\value{
json as data.frame
}
\description{
Transform influxdb json result into a data.frame
}

